+++ lettria/__init__.py
@@ -3,5 +3,6 @@
 from .client import Client
 from .NLP import NLP
 from .sentiment import Sentiment
+from .IO import load_results, load_result
 
 __all__ = ['Client', 'NLP', 'Sentiment']